Move SqlObjectBuilder to opengever.testing.
